API: move CachedDiffusionModel up to .diffusion
