+++ src/artkit/model/diffusion/base/__init__.py
@@ -18,5 +18,4 @@
 Base classes for the diffusion model
 """
 
-from ._adapter import *
 from ._diffusion import *
